src/requirements/stage.txt
# Local development dependencies go here
-r base.txt

coverage==4.4.1
flake8==3.3.0 # pyup: != 2.6.0
django-coverage-plugin==1.5.0

Sphinx==1.6.2
django-extensions==1.7.9
Werkzeug==0.15.3
django-test-plus==1.0.17
factory-boy==2.8.1

django-debug-toolbar==1.8

invoke==0.19.0

# improved REPL
ipdb==0.10.3

pytest-cov==2.5.1
pytest-django==3.1.2
pytest-sugar==0.8.0

# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.

# WSGI Handler
# ------------------------------------------------
gevent==1.4.0
gunicorn==19.9.0

# Static and Media Storage
# ------------------------------------------------
boto==2.49.0
django-storages-redux==1.3.3


# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------


# Raven is the Sentry client
# --------------------------
raven==6.10.0

# custom packages
python-dateutil
